example/requirements.txt
streamlit-audiorecorder
google-api-python-client
google-auth-httplib2
google-auth-oauthlib